Finalize Step 2 and Finish Error Middleware

+++ api/index.js
@@ -14,6 +14,16 @@ class APIError extends Error {
   }
 }
 
+
+const isInvalidString = (input) => {
+	return typeof input !== 'string' || input.trim().length === 0;
+}
+
+const isInvalidID = (input) => {
+	const ID = parseInt(input, 10);
+	return Number.isNaN(ID);
+}
+
 const SALT_ROUNDS = 10;
 
 const app = express();
@@ -37,11 +47,19 @@ const pool = new Pool({
 app.post("/auth/register", async (req, res, next) => {
 	try {
 		const { client_name, client_email, client_dob, password } = req.body;
-		//console.log(req.body);
+		var isInvalid = false;
 		if (!client_name || !client_email || !password) {
-			// return res.status(400).json({ message: "client_name, client_email, and password are required." });
 			return next(new APIError('MISSING_INFORMATION', 'Your name, email and password are required.', 400, 'One of the parameters is missing.'));
 		}
+
+		if (isInvalidString(client_name)) isInvalid = true;
+		if (isInvalidString(client_email)) isInvalid = true;
+		if (isInvalidString(password)) isInvalid = true;
+
+		if (isInvalid) {
+			return next(new APIError('INVALID_DATA', 'Your inputs include incorrect data types.', 400, 'Some of the parameters are not like the others...'))
+		}
+
 		const password_hash = await bcrypt.hash(password, SALT_ROUNDS);
 		const result = await pool.query(
 			`INSERT INTO clients (client_name, client_email, client_dob, password_hash)
@@ -61,19 +79,27 @@ app.post("/auth/login", async (req, res, next) => {
 	try {
 		const { client_email, password } = req.body;
 		if (!client_email || !password) {
-			return next(new APIError('MISSING_INFORMATION', 'Your email and password are required.', 400, 'One of the parameters is missisng. '));
+			return next(new APIError('MISSING_PARAMETER', 'Your email and password are required.', 400, 'One of the parameters is missisng.'));
 		}
+		var isInvalid = false;
+		if (isInvalidString(client_email)) isInvalid = true;
+		if (isInvalidString(password)) isInvalid = true;
+
+		if (isInvalid) {
+			return next(new APIError('INVALID_DATA', 'Your inputs include incorrect data types.', 400, 'Your parameters need to follow their correct data type.'));
+		}
+
 		const result = await pool.query(
 			`SELECT * FROM clients WHERE client_email = $1`,
 			[client_email]
 		);
 		const client = result.rows[0];
 		if (!client) {
-			return next(new APIError('INVALID_INFORMATION', 'Your email and password are invalid.', 401, 'One or both of the parameters are incorrect.'))
+			return next(new APIError('INVALID_INFORMATION', 'Your email is invalid.', 401, 'Your email does not exist.'))
 		}
 		const match = await bcrypt.compare(password, client.password_hash);
 		if (!match) {
-			return next(new APIError('INVALID_INFORMATION', 'Your email and password are invalid.', 401, 'One or both of the parameters are incorrect.'))
+			return next(new APIError('INVALID_INFORMATION', 'Your password is invalid.', 401, 'Your password does not match.'))
 		}
 		const { password_hash, ...safeClient } = client;
 		res.status(200).json({ authenticated: true, client: safeClient });
@@ -99,7 +125,14 @@ app.post("/clients", async (req, res, next) => {
 	try {
 		const { name, email } = req.body;
 		if (!name || !email) {
-			return res.status(400).json("missing parameters");
+			return next(new APIError('MISSING_PARAMETER', 'A parameter is missing.', 400, 'One of the parameters does not contain any data.'))
+		}
+		var isInvalid = false;
+		if (isInvalidString(name)) isInvalid = true;
+		if (isInvalidString(email)) isInvalid = true;
+		
+		if (isInvalid) {
+			return next(new APIError('INVALID_DATA', 'Your inputs include incorrect data types', 400, 'Please fix.'));
 		}
 		const result = await pool.query(`INSERT INTO clients (client_name, client_email)
 		VALUES ($1, $2) RETURNING client_id, client_name, client_email, client_dob`, [name, email]);
@@ -114,6 +147,10 @@ app.post("/clients", async (req, res, next) => {
 app.get("/clients/:id", async (req, res, next) => {
 	try {
 		const {id} = req.params;
+
+		if (isInvalidID(id)) {
+			return next(new APIError('INVALID_DATA', 'The ID is invalid.', 400, 'The get function is very wrong.'))
+		}
 		const result = await pool.query(`SELECT client_id, client_name, client_email, client_dob FROM clients WHERE client_id = $1`, [id]);
 		res.status(200).json(result.rows[0]);
 	} catch (e) {
@@ -126,6 +163,15 @@ app.put("/clients/:id", async (req, res, next) => {
 	try {
 		const {name, email} = req.body;
 		const {id} = req.params;
+		var isInvalid = false;
+
+		if (isInvalidString(name)) isInvalid = true;
+		if (isInvalidString(email)) isInvalid = true;
+		if (isInvalidID(id)) isInvalid = true;
+
+		if (isInvalid) {
+			return next(new APIError('INVALID_DATA', 'Your inputs include incorrect data types', 400, 'Your parameters need to follow the correct data type.'));
+		}
 
 		const result = await pool.query(`UPDATE clients
 			SET client_name = $1, client_email = $2 WHERE client_id = $3
@@ -141,6 +187,10 @@ app.put("/clients/:id", async (req, res, next) => {
 app.delete("/clients/:id", async (req, res, next) => {
 	try {
 		const {id} = req.params;
+		console.log(id);
+		if (isInvalidID(id)) {
+			return next(new APIError('INVALID_DATA', 'Your inputs include incorrect data types', 400, 'Your parameters need to follow the correct data type.'));
+		}
 		await pool.query(`DELETE FROM clients
 			WHERE client_id = $1`, [id]);
 		res.status(204).json({success: true});
@@ -168,7 +218,10 @@ app.post("/departments", async (req, res, next) => {
 	try {
 		const { name } = req.body;
 		if (!name) {
-			return res.status(400).json("missing parameters");
+			return next(new APIError('MISSING_PARAMETER', 'A parameter is missing.', 400, 'One of the parameters does not contain any data.'));
+		}
+		if (isInvalidString(name)) {
+			return next(new APIError('INVALID_DATA', 'Your inputs include incorrect data types', 400, 'Your parameters need to follow the correct data type.'));
 		}
 		const result = await pool.query(`INSERT INTO departments (dep_name)
 		VALUES ($1) RETURNING *`, [name]);
@@ -183,6 +236,9 @@ app.post("/departments", async (req, res, next) => {
 app.get("/departments/:id", async (req, res, next) => {
 	try {
 		const {id} = req.params;
+		if (isInvalidID(id)) {
+			return next(new APIError('INVALID_DATA', 'Your inputs include incorrect data types', 400, 'Your parameters need to follow the correct data type.'));
+		}
 		const result = await pool.query(`SELECT * FROM departments WHERE dep_id = $1`, [id]);
 		res.status(200).json(result.rows[0]);
 	} catch (e) {
@@ -195,6 +251,13 @@ app.put("/departments/:id", async (req, res, next) => {
 	try {
 		const {name} = req.body;
 		const {id} = req.params;
+		var isInvalid = false;
+
+		if (isInvalidString(name)) isInvalid = true;
+		if (isInvalidID(id)) isInvalid = true;
+		if (isInvalid) {
+			return next(new APIError('INVALID_DATA', 'Your inputs include incorrect data types', 400, 'Your parameters need to follow the correct data type.'));
+		}
 		const result = await pool.query(`UPDATE departments
 			SET dep_name = $1 WHERE dep_id = $2
 			RETURNING *`, [name, id]);
@@ -209,6 +272,9 @@ app.put("/departments/:id", async (req, res, next) => {
 app.delete("/departments/:id", async (req, res) => {
 	try {
 		const {id} = req.params;
+		if (isInvalidID(id)) {
+			return next(new APIError('INVALID_DATA', 'Your inputs include incorrect data types', 400, 'Your parameters need to follow the correct data type.'));
+		}
 		await pool.query(`DELETE FROM departments
 			WHERE dep_id = $1`, [id]);
 		res.status(204).json({success: true});
@@ -218,7 +284,7 @@ app.delete("/departments/:id", async (req, res) => {
 });
 
 app.use((e, req, res, next) => {
-	console.error("There is an error occuring: ", e);
+	console.error("An API error has occured: ", e);
   if (e instanceof APIError) {
     res.status(e.status).json({
       code: e.code,
